requirements_thesis.txt
amqp==2.5.1
asgiref==3.2.3
beautifulsoup4==4.7.1
billiard==3.6.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.138
botocore==1.12.138
catalogue==1.0.0
celery==4.3.0
celery-progress==0.0.5
certifi==2019.3.9
chardet==3.0.4
cloudpickle==1.1.1
cssselect==1.0.3
cycler==0.10.0
cymem==2.0.2
Cython==0.29.13
dask==1.2.2
dataclasses==0.6
de-core-news-sm==2.1.0
decorator==4.4.2
dill==0.3.0
Django==3.0.2
django-celery-results==1.1.2
django-debug-toolbar==1.11
django-emoji==2.2.2
django-guardian==2.1.0
django-npm==1.0.0
django-picklefield==2.0
djongo==1.2.32
docutils==0.14
emoji==0.5.2
en-core-web-sm==2.1.0
feedparser==5.2.1
future==0.18.0
gensim==3.7.2
gunicorn==19.9.0
html5lib==1.0.1
hurry.filesize==0.9
idna==2.8
importlib-metadata==0.23
isodate==0.6.0
jmespath==0.9.4
jsonschema==2.6.0
kiwisolver==1.1.0
kombu==4.6.5
langdetect==1.0.7
locket==0.2.0
lxml==4.3.3
matplotlib==3.1.0
more-itertools==7.2.0
murmurhash==1.0.2
mysqlclient==1.4.2.post1
networkx==2.4
nltk==3.4.1
numpy==1.16.3
oauthlib==3.1.0
pandas==0.24.2
partd==0.3.10
plac==0.9.6
preshed==2.0.1
psutil==5.6.3
pymongo==3.8.0
pyparsing==2.4.0
PySocks==1.7.1
python-dateutil==2.8.0
python-twitter==3.5
pytz==2019.3
random-user-agent==1.0.1
rdflib==4.2.2
readability-lxml==0.7.1
redis==3.2.0
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
smart-open==1.8.3
soupsieve==1.9.1
spacy==2.1.3
spacymoji==2.0.0
sqlparse==0.3.0
srsly==0.0.5
thinc==7.0.4
toolz==0.9.0
tqdm==4.31.1
tweepy==3.8.0
urllib3==1.24.2
vine==1.3.0
wasabi==0.2.1
webencodings==0.5.1
zipp==0.6.0
